Improvements to LineDrawing scenario - Add drag drawing of current line - Add undo/redo - LineCanvas is now more mutable with StraightLine now public and mutable
